Started the Project - PseudoCode and basic structure.
Initial commit for the Connect Four game code.
Using OOP C# approach.

Currently not using any animations, images or external resources, but structuring the program to receive further enhancements to achieve an exceptional UX, with unique UI.

This is a solo project and I will be the only one contributing.